refactor : BridgeGenerator, RandomBridgeGenerator 패키지 이동
